requirements.txt
Appium-Python-Client==5.1.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
colorama==0.4.6
execnet==2.1.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.1
PySocks==1.7.1
pytest==8.4.1
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-rerunfailures==15.1
pytest-xdist==3.7.0
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
timestamp==0.0.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
